Shuffle adapter classes out of models and behind protocols
