+++ jobbers/adapters/protocols.py
@@ -3,6 +3,8 @@ Protocol definitions for all pluggable adapters.
 
 Routing:
 - `RoutingBackendReadOnlyError` — raised by read-only backends on write operations.
+- `QueueConfigProtocol` — interface for queue/role configuration and refresh-tag management.
+- `TaskRoutingConfigProtocol` — interface for task routing configuration.
 - `RoutingBackendProtocol` — interface all routing backends must implement.
 
 Task storage / dead-letter queue:
@@ -32,6 +34,33 @@ if TYPE_CHECKING:
 
 class RoutingBackendReadOnlyError(Exception):
     """Raised when a write operation is attempted on a read-only routing backend."""
+
+
+@runtime_checkable
+class QueueConfigProtocol(Protocol):
+    """Interface for queue/role configuration and refresh-tag management."""
+
+    async def get_queue_config(self, queue: str) -> QueueConfig | None: ...
+    async def save_queue_config(self, queue_config: QueueConfig) -> None: ...
+    async def delete_queue(self, queue_name: str) -> None: ...
+    async def get_all_queues(self) -> list[str]: ...
+    async def get_queues(self, role: str) -> set[str]: ...
+    async def save_role(self, role: str, queues_set: set[str]) -> str: ...
+    async def get_all_roles(self) -> list[str]: ...
+    async def delete_role(self, role: str) -> None: ...
+    async def get_queue_limits(self, queues_set: set[str]) -> dict[str, int | None]: ...
+    async def get_refresh_tag(self, role: str) -> ULID: ...
+    async def bump_refresh_tag(self, role: str) -> str: ...
+    async def bump_refresh_tags_for_queue(self, queue_name: str) -> list[str]: ...
+
+
+@runtime_checkable
+class TaskRoutingConfigProtocol(Protocol):
+    """Interface for task routing configuration."""
+
+    async def get_routing_config(self, task_name: str, task_version: int) -> RoutingConfig | None: ...
+    async def save_routing_config(self, routing_config: RoutingConfig) -> None: ...
+    async def delete_routing_config(self, task_name: str, task_version: int) -> bool: ...
 
 
 @runtime_checkable
