+++ tests/models/test_task_scheduler.py
@@ -5,7 +5,8 @@ import datetime as dt
 import pytest
 import pytest_asyncio
 
-from jobbers.models.queue_config import QueueConfig, QueueConfigAdapter
+from jobbers.adapters.sql import SQLQueueConfigAdapter
+from jobbers.models.queue_config import QueueConfig
 from jobbers.models.task import Task
 from jobbers.models.task_scheduler import TaskScheduler
 from jobbers.models.task_status import TaskStatus
@@ -20,12 +21,12 @@ def make_task(task_id: str = "01JQC31AJP7TSA9X8AEP64XG08", queue: str = "default
 
 @pytest_asyncio.fixture
 async def qca(session_factory):
-    yield QueueConfigAdapter(session_factory)
+    yield SQLQueueConfigAdapter(session_factory)
 
 
 @pytest_asyncio.fixture
 async def scheduler(redis, dummy_task_adapter, session_factory):
-    qca = QueueConfigAdapter(session_factory)
+    qca = SQLQueueConfigAdapter(session_factory)
     yield TaskScheduler(redis, dummy_task_adapter, qca.get_all_queues)
